fix: 메모 추가 창에서 메모를 편집할때, 이미지를 추가했을 때, 추가된 이미지가 안 사라지는 오류 수정
