+++ src/pages/home/subPages/create/components/CreatedMemoList/CreatedMemoCard/CreatedMemoCardHeader/index.ts
@@ -1,0 +1,1 @@
+export { default as CreatedMemoCardHeader } from './CreatedMemoCardHeader';
